+++ tests/e2e/nextjs14/src/app/page.tsx
@@ -1,0 +1,11 @@
+import { init } from "../../../../../index";
+
+init({ publicKey: "" });
+
+export default function Home() {
+  return (
+    <main>
+      <div>Test</div>
+    </main>
+  );
+}