+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Version 1.1.8
+
+- improve treatment when no api key is detected
+
+# Version 1.1.7
+
+- change logError input type
+
+# Version 1.1.6
+
+- change logError return type
+
 # Version 1.1.5
 
 - log and logError no more have required extra args